update with new fields, fix Multiplier type

+++ IbFlexReader/IbFlexReader.Contracts/Ib/ComplexPosition.cs
@@ -20,7 +20,7 @@
 
         public string Issuer { get; set; }
 
-        public int? Multiplier { get; set; }
+        public double? Multiplier { get; set; }
 
         public double? Strike { get; set; }
